Increase the size of signed int to avoid overflows

+++ loom/script/native/lsLuaBridge.h
@@ -2324,7 +2324,7 @@ template<> struct Stack<
     {
         // Needs to be cast to a signed int first, then to an unsigned int
         // Casting negative real numbers to unsigned int is only defined for the range (-1, UTYPE_MAX+1)
-        return static_cast<unsigned int>(static_cast<int>(luaL_checknumber(L, index)));
+        return static_cast<unsigned int>(static_cast<int64_t>(luaL_checknumber(L, index)));
     }
 };
